some Modification s in the DB functions.

+++ ToDo-App/convex/todos.js
@@ -30,9 +30,13 @@ export const toggleTodos = mutation({
   },
 });
 
-export const deleteTodos = mutation({
+export const deleteTodo = mutation({
   args: { id: v.id("todos") },
   handler: async (context, args) => {
+    const todo = await context.db.get(args.id);
+    if (!todo) {
+      throw new ConvexError("No such todo is found");
+    }
     await context.db.delete(args.id);
   },
 });
@@ -40,7 +44,11 @@ export const deleteTodos = mutation({
 export const updateTodo = mutation({
   args: { id: v.id("todos"), text: v.string() },
   handler: async (context, args) => {
-    const todo = await context.db.patch(args.id, { text: args.text });
+    const todo = await context.db.get(args.id);
+    if (!todo) {
+      throw new ConvexError("No such todo is found");
+    }
+    const updatedTodo = await context.db.patch(args.id, { text: args.text });
   },
 });
